functions/src/requirements.txt
slackclient
boto3
